rename tsx to ts wherever no react is used
