+++ src/components/facebook/onLogoutRequested.ts
