+++ src/components/nav/NotFound.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { Page } from "../layouts/Page";
+import { Page } from "../page/Page";
 
 export const NotFound = () => <Page>
     <p>Not Found!</p>
