+++ src/components/volunteering/SearchBar.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { HPSearchBar } from "../search/HPSearchBar";
 import { HPSearchResult } from "../search/HPSearchResult";
 import { findVolunteeringOpportunities } from "./findVolunteeringOpportunities";
-import { Page } from "../layouts/Page"
+import { Page } from "../page/Page"
 
 export const SearchBar = (props) => {
     const { searchBarValue, onSearchBarValueChange } = props;
